Move Outpoint to primitives
